refactor(plants): move plant-edit to components/edit (step 11)
- Move plant-edit component to components/edit/plant-edit/
- Update imports for services and models to new paths
- Keep selector as 'app-plant-detail' for consistency
- Update routing configuration in app.routes.ts
- Remove old plant-edit directory
- Update plan document to mark step 11 as completed
